+++ packages/featureserver/src/field-map.js
@@ -1,0 +1,6 @@
+module.exports = {
+  string: 'esriFieldTypeString',
+  double: 'esriFieldTypeDouble',
+  integer: 'esrifieldTypeInteger',
+  date: 'esriFieldTypeDate'
+}