+++ packages/featureserver/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.3.1] - 2017-06-15
 ### Fixed
 * Serialize date fields as UNIX timestamps
 * Response when objectids query is for a single feature
@@ -125,6 +125,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2017-01-25
 * Initial Release
 
+[2.3.1]: https://github.com/featureserver/featureserver/compare/v2.3.1...v2.3.0
 [2.3.0]: https://github.com/featureserver/featureserver/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/featureserver/featureserver/compare/v2.2.0...v2.1.0
 [2.1.0]: https://github.com/featureserver/featureserver/compare/v2.0.0...v2.1.0